requirements.txt
transformers
torch
torchaudio
torchvision
sentence-transformers
tiktoken
rank-bm25
pandas
numpy
scikit-learn
selenium
webdriver-manager
beautifulsoup4
flashtext
openpyxl
schedule
requests
filelock
python-dotenv
